Implemented UnsetExtra on IScopeObserver
